# perf(crypto): use @chainsafe/blst