--- a/packages/crypto-key-pair-bls12-381/package.json
+++ b/packages/crypto-key-pair-bls12-381/package.json
@@ -24,7 +24,7 @@
 		"@mainsail/crypto-config": "workspace:*",
 		"@mainsail/kernel": "workspace:*",
 		"@mainsail/utils": "workspace:*",
-		"@noble/bls12-381": "^1.2.0",
+		"@chainsafe/blst": "^0.2.8",
 		"@scure/bip39": "^1.0.0",
 		"bls12-381-keygen": "^0.2.1",
 		"wif": "^2.0.6"