--- a/packages/crypto-key-pair-bls12-381/source/pair.ts
+++ b/packages/crypto-key-pair-bls12-381/source/pair.ts
@@ -1,6 +1,6 @@
+import { SecretKey } from "@chainsafe/blst";
 import { inject, injectable } from "@mainsail/container";
 import { Contracts, Identifiers } from "@mainsail/contracts";
-import { SecretKey } from "@chainsafe/blst";
 import { mnemonicToSeedSync } from "@scure/bip39";
 import { deriveChild, deriveMaster } from "bls12-381-keygen";
 import WIF from "wif";