--- a/packages/crypto-key-pair-bls12-381/source/public.ts
+++ b/packages/crypto-key-pair-bls12-381/source/public.ts
@@ -1,6 +1,6 @@
+import { aggregatePubkeys, PublicKey } from "@chainsafe/blst";
 import { inject, injectable } from "@mainsail/container";
 import { Contracts, Identifiers } from "@mainsail/contracts";
-import { aggregatePubkeys, PublicKey } from "@chainsafe/blst";
 
 @injectable()
 export class PublicKeyFactory implements Contracts.Crypto.IPublicKeyFactory {
@@ -16,12 +16,14 @@ export class PublicKeyFactory implements Contracts.Crypto.IPublicKeyFactory {
 	}
 
 	public async fromMultiSignatureAsset(asset: Contracts.Crypto.IMultiSignatureAsset): Promise<string> {
-		return Buffer.from(aggregatePubkeys(asset.publicKeys.map(pub => PublicKey.fromBytes(Buffer.from(pub, "hex")))).toBytes()).toString("hex");
+		return Buffer.from(
+			aggregatePubkeys(asset.publicKeys.map((pub) => PublicKey.fromBytes(Buffer.from(pub, "hex")))).toBytes(),
+		).toString("hex");
 	}
 
 	public async verify(publicKey: string): Promise<boolean> {
 		try {
-			PublicKey.fromBytes(Buffer.from(publicKey, "hex")).keyValidate()
+			PublicKey.fromBytes(Buffer.from(publicKey, "hex")).keyValidate();
 		} catch {
 			return false;
 		}
@@ -30,6 +32,8 @@ export class PublicKeyFactory implements Contracts.Crypto.IPublicKeyFactory {
 	}
 
 	public async aggregate(publicKeys: Buffer[]): Promise<string> {
-		return Buffer.from(aggregatePubkeys(publicKeys.map(pub => PublicKey.fromBytes(pub))).toBytes()).toString("hex");
+		return Buffer.from(aggregatePubkeys(publicKeys.map((pub) => PublicKey.fromBytes(pub))).toBytes()).toString(
+			"hex",
+		);
 	}
 }