--- a/packages/crypto-signature-bls12-381/package.json
+++ b/packages/crypto-signature-bls12-381/package.json
@@ -23,7 +23,7 @@
 		"@mainsail/contracts": "workspace:*",
 		"@mainsail/kernel": "workspace:*",
 		"@mainsail/utils": "workspace:*",
-		"@noble/bls12-381": "^1.2.0"
+		"@chainsafe/blst": "^0.2.8"
 	},
 	"devDependencies": {
 		"uvu": "^0.5.6"