--- a/packages/crypto-signature-bls12-381/package.json
+++ b/packages/crypto-signature-bls12-381/package.json
@@ -19,11 +19,11 @@
 		"test:file": "uvu -r tsm source"
 	},
 	"dependencies": {
+		"@chainsafe/blst": "^0.2.8",
 		"@mainsail/container": "workspace:*",
 		"@mainsail/contracts": "workspace:*",
 		"@mainsail/kernel": "workspace:*",
-		"@mainsail/utils": "workspace:*",
-		"@chainsafe/blst": "^0.2.8"
+		"@mainsail/utils": "workspace:*"
 	},
 	"devDependencies": {
 		"uvu": "^0.5.6"