--- a/packages/crypto-signature-bls12-381/source/signature.ts
+++ b/packages/crypto-signature-bls12-381/source/signature.ts
@@ -1,16 +1,16 @@
 import { injectable } from "@mainsail/container";
 import { Contracts } from "@mainsail/contracts";
 import { ByteBuffer } from "@mainsail/utils";
-import { aggregateSignatures, sign, verify } from "@noble/bls12-381";
+import { aggregateSignatures, verify, PublicKey, SecretKey, Signature as Sig } from "@chainsafe/blst";
 
 @injectable()
 export class Signature implements Contracts.Crypto.ISignature {
 	public async sign(message: Buffer, privateKey: Buffer): Promise<string> {
-		return Buffer.from(await sign(message, privateKey)).toString("hex");
+		return Buffer.from(SecretKey.fromBytes(privateKey).sign(message).toBytes()).toString("hex");
 	}
 
 	public async verify(signature: Buffer, message: Buffer, publicKey: Buffer): Promise<boolean> {
-		return verify(signature, message, publicKey);
+		return verify(message, PublicKey.fromBytes(publicKey), Sig.fromBytes(signature));
 	}
 
 	public serialize(buffer: ByteBuffer, signature: string): void {
@@ -22,6 +22,6 @@ export class Signature implements Contracts.Crypto.ISignature {
 	}
 
 	public async aggregate(signatures: Buffer[]): Promise<string> {
-		return Buffer.from(aggregateSignatures(signatures)).toString("hex");
+		return Buffer.from(aggregateSignatures(signatures.map(s => Sig.fromBytes(s))).toBytes()).toString("hex");
 	}
 }