<!--
Thanks for your interest in the project. Bugs filed and PRs submitted are appreciated!

Please make sure you're familiar with and follow the instructions in the [contributing guidelines](https://ark.dev/docs/program-incentives/guidelines/contributing).

Please fill out the information below to expedite the review and (hopefully) merge of your pull request!
-->

## Summary

<!-- What changes are being made? -->

Replace `@noble/bls12-381` with https://github.com/ChainSafe/blst-ts

before:
```
verify-bls12-381 x 28 ops/sec @ 34ms/op
```

after

```
verify-bls12-381 x 846 ops/sec @ 1ms/op
```

it's still approx. 10 slower than schnorr, but it's something we can work with.

```
verify-schnorr x 8,772 ops/sec @ 113μs/op
```

<!-- Why are these changes necessary? -->

<!-- How were these changes implemented? -->

## Checklist

<!-- Have you done all of these things?  -->

- [ ] Documentation _(if necessary)_
- [ ] Tests _(if necessary)_
- [ ] Ready to be merged

<!-- Feel free to add additional comments. -->